{{cookiecutter.project_name}}/requirements.txt
bootstrap-flask>=2.0.2
Flask>=2.1.3
Flask-Bcrypt>=1.0.1
Flask-DB>=0.3.2
Flask-Login>=0.6.1
Flask-Mail>=0.9.1
Flask-Moment>=1.0.4
Flask-SQLAlchemy>=2.5.1
Flask-WTF[email]>=1.0.1
PyJWT>=2.4.0
python-dotenv>=0.20.0
werkzeug==2.1.2
